+++ src/entity/User.java
@@ -63,7 +63,13 @@ public interface User {
 
     String getPaceType();
 
-    int getRequiredCalories(); // Should default to 0
+
+//    HashMap<String, HashMap<String, Boolean>> weightPaceType();
+//    // {loseWeight : {typical : False, moderate: True, extreme: False}
+//
+//    HashMap<String, String> getPaceType();
+
+    double getRequiredCalories(); // Should default to 0
 
     // setters
     void setPassword(String password);
@@ -92,6 +98,8 @@ public interface User {
 
     void setPaceType(String paceType);
 
-    void setRequiredCalories(int requiredCalories);
+
+    void setRequiredCalories(double requiredCalories);
+
 
 }
